complete standalone mode
