+++ connectors/rocketmq-connect-jdbc/src/main/java/org/apache/rocketmq/connect/jdbc/connector/JdbcSourceConfig.java
@@ -207,7 +207,7 @@ public class JdbcSourceConfig extends AbstractConfig {
     private String incrementingColumnName;
     private List<String> timestampColumnNames;
     private long timestampDelayIntervalMs;
-    private Long timestampInitial;
+    private Long timestampInitial = TIMESTAMP_INITIAL_DEFAULT;
     private Set<String> tableWhitelist;
     private Set<String> tableBlacklist;
     private String schemaPattern;
@@ -231,7 +231,10 @@ public class JdbcSourceConfig extends AbstractConfig {
         this.incrementingColumnName = config.getString(INCREMENTING_COLUMN_NAME_CONFIG);
         this.timestampColumnNames = getList(config, TIMESTAMP_COLUMN_NAME_CONFIG);
         timestampDelayIntervalMs = config.getLong(TIMESTAMP_DELAY_INTERVAL_MS_CONFIG);
-        this.timestampInitial = config.containsKey(TIMESTAMP_INITIAL_CONFIG) ? config.getLong(TIMESTAMP_INITIAL_CONFIG) : TIMESTAMP_INITIAL_DEFAULT;
+//        this.timestampInitial=config.getLong(TIMESTAMP_INITIAL_CONFIG,TIMESTAMP_INITIAL_DEFAULT);
+        if (config.containsKey(TIMESTAMP_INITIAL_CONFIG)){
+            this.timestampInitial=config.getLong(TIMESTAMP_INITIAL_CONFIG);
+        }
         this.tableWhitelist = new HashSet<>(getList(config, TABLE_WHITELIST_CONFIG));
         this.tableBlacklist = new HashSet<>(getList(config, TABLE_BLACKLIST_CONFIG));
         this.schemaPattern = config.getString(SCHEMA_PATTERN_CONFIG);
