+++ rocketmq-connect-runtime/connect-distributed.sh
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 echo "run rumtime worker"
-cd target/distribution/ && java -cp .:./conf/:./lib/* org.apache.rocketmq.connect.runtime.ConnectStartup -c conf/connect.conf
+cd target/distribution/ && java -cp .:./conf/:./lib/* org.apache.rocketmq.connect.runtime.DistributedConnectStartup -c conf/connect.conf
