+++ rocketmq-connect-runtime/src/main/java/org/apache/rocketmq/connect/runtime/service/ConfigManagementService.java
@@ -18,8 +18,10 @@
 package org.apache.rocketmq.connect.runtime.service;
 
 import io.openmessaging.connector.api.component.connector.Connector;
+
 import java.util.List;
 import java.util.Map;
+
 import org.apache.rocketmq.connect.runtime.common.ConnectKeyValue;
 import org.apache.rocketmq.connect.runtime.config.ConnectConfig;
 import org.apache.rocketmq.connect.runtime.utils.Plugin;
@@ -29,7 +31,6 @@ import org.apache.rocketmq.connect.runtime.utils.Plugin;
  * same configs.
  */
 public interface ConfigManagementService {
-
 
 
     /**
@@ -44,11 +45,12 @@ public interface ConfigManagementService {
 
     /**
      * Configure class with the given key-value pairs
+     *
      * @param config can be DistributedConfig or StandaloneConfig
      */
-     default void configure(ConnectConfig config){
+    default void configure(ConnectConfig config) {
 
-     }
+    }
 
     /**
      * Get all connector configs from the cluster filter out DELETE set to 1
