+++ rocketmq-connect-runtime/src/main/java/org/apache/rocketmq/connect/runtime/service/memory/MemoryConfigManagementServiceImpl.java
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package org.apache.rocketmq.connect.runtime.service.memory;
 
 
@@ -64,18 +80,17 @@ public class MemoryConfigManagementServiceImpl implements ConfigManagementServic
     public void start() {
         connectorKeyValueStore.load();
         taskKeyValueStore.load();
-//        sendOnlineConfig();
     }
 
     @Override
     public void stop() {
-//        sendSynchronizeConfig();
         connectorKeyValueStore.persist();
         taskKeyValueStore.persist();
     }
 
     /**
      * get all connector configs enabled
+     *
      * @return
      */
     @Override
@@ -94,6 +109,7 @@ public class MemoryConfigManagementServiceImpl implements ConfigManagementServic
 
     /**
      * get all connector configs include deleted
+     *
      * @return
      */
     @Override
@@ -138,7 +154,6 @@ public class MemoryConfigManagementServiceImpl implements ConfigManagementServic
         }
         final Connector connector = (Connector) clazz.getDeclaredConstructor().newInstance();
         connector.validate(configs);
-        connector.init(configs);
         connectorKeyValueStore.put(connectorName, configs);
         recomputeTaskConfigs(connectorName, connector, currentTimestamp, configs);
         return "";
@@ -221,7 +236,7 @@ public class MemoryConfigManagementServiceImpl implements ConfigManagementServic
 
     @Override
     public void registerListener(ConnectorConfigUpdateListener listener) {
-        this.connectorConfigUpdateListener=listener;
+        this.connectorConfigUpdateListener = listener;
     }
 
     private void triggerListener() {
