+++ rocketmq-connect-runtime/src/main/resources/connect-standalone.conf
@@ -13,10 +13,8 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-workerId=DEFAULT_WORKER_1
-storePathRootDir=/xxx/storeRoot
-## stand alone offset storage
-offset.storage.file.filename=/tmp/connect.offsets
+workerId=standalone-worker
+storePathRootDir=/tmp/storeRoot
 
 ## Http port for user to access REST API
 httpPort=8082
@@ -34,6 +32,5 @@ clusterName="DefaultCluster"
 
 # Source or sink connector jar file dir,The default value is rocketmq-connect-sample
 pluginPaths=/xxx/connector-plugins
-
